✨ npm-publish (#2)
Creates `npm-publish~v1.yaml` workflow that publishes a package to the
npm registry. Extends `npm-ci-build` to pack the package and publish it
as a workflow artifact (for the other job to download it and publish to
npm).

Removes tabs from .editorconfig because (a) it caused issues with YAML in
markdown and (b) this only has YAML files which disallows tabs.